requirements/common.txt
Click==7.0
Python-Dotenv==0.10.*
Flask==1.1.*
Flask-Bootstrap==3.3.*
Flask-Caching==1.7.*
Flask-Cors==3.0.*
Flask-Env==2.0.*
Flask-Login==0.4.*
Flask-Mail==0.9.*
Flask-Migrate==2.5.*
Flask-SQLAlchemy==2.4.*
Flask-WTF==0.14.*
lxml==4.4.*
openpyxl==3.0.*
Pillow==6.2.*
psycopg2==2.8.*
pytz==2019.*
requests==2.22.*
WeasyPrint==50
xlrd==1.2.*